requirements.txt
# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.1
pydantic-settings>=2.3.0

# LLM
anthropic>=0.28.0

# Vector DB
qdrant-client>=1.9.1

# Embeddings — Python 3.13 compatible
torch>=2.9.0
sentence-transformers>=3.3.0

# Sparse search
rank-bm25>=0.2.2

# Queue / Workers
celery>=5.4.0
redis>=5.0.4

# Data
pandas>=2.2.2
numpy>=1.26.4
tqdm>=4.66.4

# Document ingestion
PyPDF2>=3.0.1
python-docx>=1.1.2
langchain-text-splitters>=0.2.1

# HTTP
httpx>=0.27.0
aiofiles>=23.2.1

# Logging
structlog>=24.2.0
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.1
tenacity>=8.3.0
xxhash>=3.4.1
python-multipart>=0.0.9
